Replaces APC specific classes with generic class
Rather than using my own, homegrown caching adapter and specify APC as
the caching mechanism in the FlickrServiceApc class, I've created a more
generic FlickrServiceCache class that accepts an instance of
\Zend\Cache\Storage\Adapter\AbstractAdapter.

While the caching can now be generic, I intend to use filesystem caching as
the caching mechanism.  I've added a tmp directory to that end.